partialPivLU fixed. some changes to HODLR_Tree.cpp

+++ src/lowRank.cpp
@@ -871,17 +871,38 @@ int PS_PseudoInverse(Eigen::MatrixXd & colMatrix,Eigen::MatrixXd & rowMatrix, Ei
     Eigen::FullPivLU<Eigen::MatrixXd> lu(tempK);
     lu.setThreshold(tol);
     rank = lu.rank();
-    int largestPivot = std::abs((lu.matrixLU())(0,0));
+    //debug code
+    //std::cout<<"numRowsSelect = "<<numRowsSelect<<" "<<"numColsSelect = "<<numColsSelect<<" rank  = "<<rank<<std::endl;
+    //std::cout<<"Rank Info  "<<(lu.matrixLU())(0,0)<<" "<<(lu.matrixLU())(rank-1,rank-1)<<std::endl;
+    //std::cout<<"Normm check "<<colMatrix.norm() - rowMatrix.norm()<<std::endl;
+    //if (std::abs((lu.matrixLU())(rank+1,rank+1))>tol){
+    //  std::cout<<"newRank "<<(lu.matrixLU())(rank+1,rank+1)<<std::endl;
+    //  rank ++;
+    //}
+    
+    //int largestPivot = std::abs((lu.matrixLU())(0,0));
+    double largestPivot = std::abs((lu.matrixLU())(0,0));
+    
     if ((rank > 0) && (largestPivot >= 1e-6)){
       V = ((lu.permutationP() * rowMatrix.transpose()).transpose()).leftCols(rank);
       Eigen::MatrixXd L_Soln = lu.matrixLU().topLeftCorner(rank,rank).triangularView<Eigen::UnitLower>().solve(V.transpose());
       V = lu.matrixLU().topLeftCorner(rank,rank).triangularView<Eigen::Upper>().solve(L_Soln).transpose();
       U = (colMatrix * lu.permutationQ()).leftCols(rank);
     }else{
+      //debug code
+      //std::cout<<"shouldnt be here"<<std::endl;
+      //
       U = Eigen::MatrixXd::Zero(colMatrix.rows(),1);
       V = Eigen::MatrixXd::Zero(rowMatrix.rows(),1);
       rank = 1;
     }
+    //debug code
+    //std::cout<<"Rank Info  "<<(lu.matrixLU())(0,0)<<" "<<(lu.matrixLU())(rank-1,rank-1)<<std::endl;
+    //std::cout<<"cond info  "<<largestPivot<<" "<<"cond "<<(largestPivot >= 1e-6)<<std::endl;
+    //std::cout<<"Entr Norm "<<colMatrix.norm() - rowMatrix.norm()<<std::endl;
+    //std::cout<<"exit norm "<<colMatrix.norm() - (U*V.transpose()).norm()<<std::endl;
+    //std::cout<<"**************"<<std::endl;
+    //
   }else{
     std::cout<<"Error! Unknown operation mode"<<std::endl;
     exit(EXIT_FAILURE);
